OXC Support: Add `resolveIndexedLocations` helper
`oxc-parser` emits AST nodes that do not contain `loc` fields, but
rather `start` and `end` indicies.

Unfortunately, that means we cannot add support for those by default, as
due to the sourcemap specification being from the depths of hell, we
must provide a line and column number, and thus must have the original
source code to convert from the index to line-and-column.